PC bootstrapping fix

+++ src/main.py
@@ -151,7 +151,7 @@ def main(cfg: DictConfig):
                 if cfg.do.bootstrap:
                     pred_adj_matrix = np.zeros((cfg.graph.num_vars, cfg.graph.num_vars))
                     for _ in range(10):
-                        indices = np.random.choice(len(synth_dataset), size=1/3 * len(synth_dataset), replace=False)
+                        indices = np.random.choice(len(synth_dataset), size=int(1/3 * len(synth_dataset)), replace=False)
                         sub_dataset = data.PartitionData(features=synth_dataset.features[indices, :-1],
                                                          targets=synth_dataset[indices].targets)
                         sub_dataset.memberships = synth_dataset.memberships[indices]
